requirements.txt
bcrypt
flask-collect
flask-jwt-extended
flask-rest-api
flask-security
gunicorn
marshmallow
peewee-migrate
psycopg2
unidecode
